+++ lib/styling/consumerTheme.types.ts
@@ -1,5 +1,9 @@
 import { CSSObject, Theme } from '@emotion/react';
 
 export type ConsumerThemeTypes = Theme & {
-  button?: (params: { kind: string; size: string }) => CSSObject;
+  button?: (params: {
+    kind?: string;
+    size?: string;
+    iconOnly?: boolean;
+  }) => CSSObject;
 };