+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## [0.0.32] - 2024-23-09
+
+### Added
+
+- Rudimentary icon support for `<Button>`
+- "Close" functionality in `<Modal>`
+
+### Changed
+
+- Base styles on `<ListItem>`
+- General `baseTheme` changes
+
+### Removed
+
+- ## N/A
+
+### Fixed
+
+- ## N/A
+
+<br/>
+
 ## [0.0.30] - 2024-17-09
 
 ### Added